### 2019-05-22 #1471 @whedon
[PRE REVIEW]: cde - R package to retrieve data from the Environment Agency Catchment Data Explorer site
**Submitting author:** @robbriers (<a href="http://orcid.org/0000-0003-0341-1203">Robert Briers</a>)
**Repository:** <a href="https://github.com/ropensci/cde" target ="_blank">https://github.com/ropensci/cde</a>
**Version:** 0.4.0
