### 2019-05-22 #1471 @danielskatz
@robbriers - I've made a PR that I think will fix the build problem: https://github.com/ropensci/cde/pull/22

Please accept it, then enter ```@whedon generate pdf``` as a new comment in this issue - If more is needed, free free to make additional fixes...
