### 2019-05-23 a1fa7b3
minor changes/fixes to JOSS paper
