+++ paper/paper.md
@@ -17,13 +17,13 @@ bibliography: paper.bib
 
 # Summary
 
-Globally, issues around water quality and quantity are expected to increase in coming decades, set against a background of already widespread degradation [@WWAPUnitedNationsWorldWaterAssessmentProgramme/UN-Water2018]. Within Europe, the [Water Framework Directive](http://ec.europa.eu/environment/water/water-framework/index_en.html) set EU-wide standards for how the quality of surface- and ground-waters across Europe is assessed and classified [@Communities2000]. Assessment of quality using the WFD is based on a range of elements that vary depending on the type of water being assessed. The elements cover biological, chemical and hydromorphological/quantitative components with a hierarchical structure. These are combined to give an overall classification of waterbodies into five classes(High, Good, Moderate, Poor and Bad) for surface waters and two classes (Good or Poor) for groundwaters. The overall aim of the WFD is that all European surface and ground waters will reach at least Good status by 2027, although there have been a number of issues with implementation [@Voulvoulis2017].
+Globally, issues around water quality and quantity are expected to increase in coming decades, set against a background of already widespread degradation [@WWAPUnitedNationsWorldWaterAssessmentProgramme/UN-Water2018]. Within Europe, the [Water Framework Directive](http://ec.europa.eu/environment/water/water-framework/index_en.html) (WFD) set EU-wide standards for how the quality of surface- and ground-waters across Europe is assessed and classified [@Communities2000]. Assessment of quality using the WFD is based on a range of elements that vary depending on the type of water being assessed. The elements cover biological, chemical and hydromorphological/quantitative components with a hierarchical structure. These are combined to give an overall classification of waterbodies into five classes (High, Good, Moderate, Poor and Bad) for surface waters and two classes (Good or Poor) for groundwaters. The overall aim of the WFD is that all European surface and ground waters will reach at least Good status by 2027, although there have been a number of issues with implementation [@Voulvoulis2017].
 
-In the UK the Environment Agency (EA) is the competent authority responsible for monitoring and assessment of water quality within England. The EA have made the reporting data relating to the requirements of the WFD available via the Catchment Data Explorer (CDE) website [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The ``cde`` package for R [@RCoreTeam2019] provides functions that facilitate the querying, download and plotting of the data available on the CDE site, allowing efficient collation of data for the analysis of trends or patterns in water quality and pressures on waterbodies across England.
+In the UK, the Environment Agency (EA) is the competent authority responsible for monitoring and assessment of water quality within England. The EA have made the reporting data relating to the requirements of the WFD available via the Catchment Data Explorer (CDE) website [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The ``cde`` package for R [@RCoreTeam2019] provides functions that facilitate the querying, download and plotting of the data available on the CDE site, allowing efficient collation of data for the analysis of trends or patterns in water quality and pressures on waterbodies across England.
 
-The package contains functions to allow users to search for waterbodies, catchments or River Basin Districts that match given search strings. Having identified the relevant sites, the following types of data can be downloaded:
+The package allows users to search for waterbodies, catchments or River Basin Districts that match given search strings. Having identified the relevant sites, the following types of data can be downloaded:
 
-*Status classification data*: either for overall water body classification or at a range of more detailed levels relating to specific quality elements (e.g. ecological, chemical or priority substances). These can be downloaded for specified year ranges and for specific waterbody types (such as lakes).
[…]
