### 2019-06-20 ropensci/cde#23 @jhollist
JOSS Review:  Issues to address
Very nice job on the package!  I am a big fan of anything that provides direct access, in R, to these kinds of data sources.  There are few things to need to be addressed in the JOSS submission, https://github.com/openjournals/joss-reviews/issues/1473, prior to publication.  These are listed below.

- Software Paper, second paragraph of intro, `cde` got split at the EOL.  Probably a rendering thing, but would be best to make sure that doesn't happen.

- Software Paper, Statement of need: Set up is there, but a clear indication as to why an R package is needed to access the Catchment Data Explorer.  A sentence or two in the second paragraph that expands on the need for the `cde` package should be added.  I buy into the need for these types of data access packages (e.g. reproducibility, automation, etc.), just need to see it in writing in the paper.

- Documentation, Statement of need: Take a look at <https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need> for details on what should be included.  Similar to the "statement of need" for the Software Paper, a senetence of two should be added to the README that lay out the reason for developing `cde`.

- Documentation, Community Guidelines:  Wasn't clear from the repo how one should contribute/report bugs, etc.  I'd suggest adding a CONTRIBUTING.md and/or adding details about this to the README

- Documentation, Example Usage:  There are two separate links to documentation sites:
  - <https://docs.ropensci.org/cde>
  - <https://ropensci.github.io/cde>
For the most part the appear very similar.  I did notice that the first plotting example in the github.io site plots a different plot than the one included in the documentation.  Make sure you link to a single site.  It may also be useful to provide a section in the README with a link to the documentation sites and a bit of text explaining how and where to find these examples
