### 2019-06-20 #1473 @jhollist
Thanks all for giving me the extra time to get this review done!  And kudos to @robbriers for a nicely done package.

I have updated the checklist above.  For each of the items I did not check I have a few things that I feel should be addressed prior to signing off on those.  I have outlined those in an issue on the package repository.    See those at https://github.com/ropensci/cde/issues/23

Any questions, please let me know.
