### 2019-06-24 9c8a713
fix for bug in year subsetting - accidentally deleted part to subset across year range
