+++ NEWS.md
@@ -1,3 +1,5 @@
+# cde (development version)
+
 cde 0.4.0 (2019-05-17)
 =========================
 
