### 2019-06-24 9010e16
updating JOSS paper, readme and index to add statement of need and links to rOpenSci docs site
