+++ README.Rmd
@@ -14,6 +14,7 @@ knitr::opts_chunk$set(
 
 # cde <img src="docs/logo/logo.png" align="right" height=140/>
 [![Travis-CI Build Status](https://travis-ci.org/ropensci/cde.svg?branch=master)](https://travis-ci.org/ropensci/cde) [![Coverage Status](https://coveralls.io/repos/github/ropensci/cde/badge.svg?branch=master)](https://coveralls.io/github/ropensci/cde?branch=master) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![](https://badges.ropensci.org/284_status.svg)](https://github.com/ropensci/onboarding/issues/284) [![DOI](https://zenodo.org/badge/92712854.svg)](https://zenodo.org/badge/latestdoi/92712854)
+[![status](http://joss.theoj.org/papers/0d35f75e861fcf47556d70571e226589/status.svg)](http://joss.theoj.org/papers/0d35f75e861fcf47556d70571e226589)
 
 ## Introduction
 
@@ -21,7 +22,7 @@ Within Europe, the [Water Framework Directive](http://ec.europa.eu/environment/w
 
 In the UK the Environment Agency (EA) is the competent authority responsible for monitoring and assessment of water quality within England. The EA have made the reporting data relating to the requirements of the WFD available via the Catchment Data Explorer (CDE) website, [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). 
 
-`cde` is a package for R which facilitates searching and download of the WFD reporting data for all waterbodies from the EA CDE website.
+`cde` is a package for R which facilitates searching and download of the WFD reporting data for all waterbodies from the EA CDE website. The ability to access these data from within the R environment allows for efficient collation and interrogation of data and reproducible  analysis of trends or patterns in water quality and pressures on waterbodies across England.
 
 The types of data that can be downloaded are: WFD status classification data, Reasons for Not Achieving Good (RNAG) status, objectives set for waterbodies, measures put in place to improve water quality and details of associated protected areas.
 
@@ -38,7 +39,7 @@ remotes::install_github("ropensci/cde")
 
 ## Basic usage
 
-Details of how to use the package can be found at [https://ropensci.github.io/cde](https://ropensci.github.io/cde).
+Details of how to use the package can be found at [https://docs.ropensci.org/cde](https://docs.ropensci.org/cde).
 
 [![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
 

[…]
