### 2019-06-25 95e5f8e
updates to readme and index. still need to do full rebuild of docs via pkgdown.
