+++ README.Rmd
@@ -22,7 +22,7 @@ Within Europe, the [Water Framework Directive](http://ec.europa.eu/environment/w
 
 In the UK the Environment Agency (EA) is the competent authority responsible for monitoring and assessment of water quality within England. The EA have made the reporting data relating to the requirements of the WFD available via the Catchment Data Explorer (CDE) website, [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). 
 
-`cde` is a package for R which facilitates searching and download of the WFD reporting data for all waterbodies from the EA CDE website. The ability to access these data from within the R environment allows for efficient collation and interrogation of data and reproducible  analysis of trends or patterns in water quality and pressures on waterbodies across England.
+`cde` is a package for R which facilitates searching and download of the WFD reporting data for all waterbodies from the EA CDE website. The ability to access these data from within the R environment allows for efficient collation and interrogation of data and reproducible analysis of trends or patterns in water quality and pressures on waterbodies across England.
 
 The types of data that can be downloaded are: WFD status classification data, Reasons for Not Achieving Good (RNAG) status, objectives set for waterbodies, measures put in place to improve water quality and details of associated protected areas.
 
@@ -40,6 +40,10 @@ remotes::install_github("ropensci/cde")
 ## Basic usage
 
 Details of how to use the package can be found at [https://docs.ropensci.org/cde](https://docs.ropensci.org/cde).
+
+## Contributing
+
+For details of how to contribute to this package, see [here](https://docs.ropensci.org/cde/CONTRIBUTING.html).
 
 [![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
 

[…]
