### 2019-06-25 85b9594
fixing intro deader on index.md
