### 2019-06-25 5c7c5e2
updating man pages to fix links to new docs site, plus other typos etc.
