+++ R/cde-class_and_methods.r
@@ -119,9 +119,9 @@ trunc_char <- function(x, col_name_lengths){
 #' 
 #' @param ... Other arguments passed on to individual methods. The only other
 #' argument implemented at present is \code{scheme}. For \code{status} and 
-#' \code{objectives} this defines which colour scheme to use with plots.It 
+#' \code{objectives} data this defines which colour scheme to use with plots. It 
 #' defaults to a viridis-based scheme (\code{scheme="vir"}). Alternatively, the 
-#' colours specified in the WFD document ca be used by specifying 
+#' colours specified in the WFD document can be used by specifying 
 #' \code{scheme="wfd"}.
 #'
 #' @importFrom graphics barplot
