+++ R/get_objectives.r
@@ -56,7 +56,7 @@
 #' @return An object of class \code{cde_df} containing the details 
 #' of the objectives set for the specified set of waterbodies.
 #' For details of the meaning of the the different columns returned, 
-#' see \url{https://ropensci.github.io/cde/articles/cde-output-reference.html}.
+#' see \url{https://docs.ropensci.org/cde/articles/cde-output-reference.html}.
 #'
 #' @export get_objectives
 #'
