### 2019-06-25 a7b915e
updating readme and index to fix image issue in rotemplate
