+++ index.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# cde <img src="docs/logo/logo.png" align="right" height=140/>
+# cde <img src="https://ropensci.github.io/cde/logo/logo.png" align="right" height=140/>
 
 [![Travis-CI Build
 Status](https://travis-ci.org/ropensci/cde.svg?branch=master)](https://travis-ci.org/ropensci/cde)
