### 2019-06-25 c15f429
removing dplyr paragraph from vignette
