### 2019-06-26 247fa1f
removing rOpensci footer from index and -devel from travis due to current issue with errors on travis
