+++ .travis.yml
@@ -12,7 +12,6 @@ os:
 - osx
 
 r:
-- devel
 - release
 - oldrel
 
