### 2019-06-26 ropensci/cde#23 @karthik
> For the most part the appear very similar. I did notice that the first plotting example in the github.io site plots a different plot than the one included in the documentation. Make sure you link to a single site. It may also be useful to provide a section in the README with a link to the documentation sites and a bit of text explaining how and where to find these examples

rOpenSci now automatically deploys the docs for all projects. So the first one should be used, not the second.
