### 2019-06-26 ropensci/cde#23 @karthik
I just realized that this package has [already been through rOpenSci's software review](https://github.com/ropensci/software-review/issues/284), so I'm glad most of these comments are about the paper.

> Documentation, Community Guidelines: Wasn't clear from the repo how one should contribute/report bugs, etc. I'd suggest adding a CONTRIBUTING.md and/or adding details about this to the README

Can be fixed with `usethis::use_tidy_contributing()`
