### 2019-06-26 ropensci/cde#23 @karthik
cc openjournals/joss-reviews#1473
