### 2019-06-29 bb9b083
updating statement of need throughout and rebooting -devel build to check
