### 2019-06-29 ropensci/cde#23 @robbriers
OK, been through the comments:

- Software paper: fixed split of 'cde' in paragraph and expanded on paragraph including Statement of need.
- Documentation: included same text as for software paper to expand on the need for the package. Also added Contribution page and link in README.
- Documentation: examples of usage. The new [docs.ropensci.org/cde](https://docs.ropensci.org/cde/) site came online after the submission, hence the confusion of the two docs sites. All links have been updated to the docs site now. The difference between the figures has also been fixed - this was due to a bug in one of the functions which has been fixed (0.4.1 Github release).
Hope this is OK, but let me know if anything else needs attention.
