### 2019-06-29 #1473 @robbriers
Updated pdf above and addressed points raised in [https://github.com/ropensci/cde/issues/23](https://github.com/ropensci/cde/issues/23).
Let me know if anything else needs attention
