### 2019-07-01 #1473 @jhollist
> Thanks for the updates @robbriers. @jhollist When you get a chance, can you check out the statement of need?

Just took a look at the new proof and statement of need looks good.  Going to check on the other things but anticipate signing off on these momentarily.
