### 2019-07-01 #1473 @jhollist
Looks good to me!  Just signed off on the last remaining items in the checklist.  👍 👍
