### 2019-07-02 5700fc8
fixing CRAN gotchas picked up by rhub
