+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cde
 Type: Package
-Title: Download Water Framework Directive (WFD) data from the Environment Agency Catchment Data Explorer (CDE) website
+Title: Download Water Framework Directive (WFD) Data from the Environment Agency Catchment Data Explorer (CDE) Website
 Version: 0.4.1
 Authors@R: person("Rob", "Briers", email = "r.briers@napier.ac.uk",
   role = c("aut", "cre"), comment = c(ORCID = "0000-0003-0341-1203"))
@@ -10,9 +10,9 @@ Description: The cde package facilitates searching and download of the
   (which can also be plotted), Reasons for Not Achieving Good (RNAG) status,
   objectives set for waterbodies, measures put in place to improve water 
   quality and details of associated protected areas. The site accessed is 
-  https://environment.data.gov.uk/catchment-planning/. The data are made 
+  <https://environment.data.gov.uk/catchment-planning/>. The data are made 
   available under the Open Government Licence v3.0 
-  (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/).
+  <https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/>.
   This package has been peer-reviewed by rOpenSci (v. 0.4.0).
 Depends:
     R (>= 3.0.0)
