### 2019-07-04 #1473 @labarba
👋  @robbriers — Please check that the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
