### 2019-07-04 #1473 @robbriers
@labarba fixed zenodo metadata as requested
