### 2019-07-04 #1473 @labarba
While you are there, add a comma after "e.g." on page 2.
