### 2019-07-04 298a3e1
fixing paper citation issues
