+++ paper/paper.bib
@@ -7,7 +7,7 @@ volume = {L 327},
 year = {2000}
 }
 
-@techreport{WWAPUnitedNationsWorldWaterAssessmentProgramme/UN-Water2018,
+@techreport{WWAPUnitedNationsWorldWaterAssessmentProgramme2018,
 address = {Paris},
 author = {{WWAP (United Nations World Water Assessment Programme)/UN-Water}},
 institution = {UNSECO},
