+++ paper/paper.bib
@@ -9,7 +9,7 @@ year = {2000}
 
 @techreport{WWAPUnitedNationsWorldWaterAssessmentProgramme2018,
 address = {Paris},
-author = {{WWAP (United Nations World Water Assessment Programme)/UN-Water}},
+author = {{WWAP - United Nations World Water Assessment Programme/UN-Water}},
 institution = {UNSECO},
 isbn = {9789231002649},
 title = {{The United Nations World Water Development Report 2018: Nature-Based Solutions for Water}},
