### 2019-07-04 #1473 @labarba
Congrats @robbriers! Your JOSS paper is accepted.

Big thanks to our editor: @karthik and reviewer: @jhollist
