### 2019-07-04 #1473 @labarba
(Just published this from my airplane seat before doors close for a long-haul. Cool.)
