### 2019-07-04 8ad2449
prepping cran submission and reactivating -devel on travis
