+++ .travis.yml
@@ -14,6 +14,7 @@ os:
 r:
 - release
 - oldrel
+- devel
 
 after_success:
   - Rscript -e 'covr::coveralls()'
